changes in counting median and average

+++ lib/ui/ui_models/estimated_task_info.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 class EstimatedTaskInfo {
   final String taskId;
-  final double average;
+  final int average;
   final int median;
 
   const EstimatedTaskInfo({@required this.taskId, this.average, this.median});
